Fix import from moving file
