+++ src/operations/extract.ts
@@ -1,5 +1,5 @@
 import { GraphSnapshot } from '../GraphSnapshot';
-import { Serializeable } from '../primitive';
+import { Serializeable } from '../schema';
 
 export function extract(data: GraphSnapshot): Serializeable.GraphSnapshot {
   return { };
