+++ src/operations/restore.ts
@@ -1,5 +1,5 @@
 import { GraphSnapshot } from '../GraphSnapshot';
-import { Serializeable } from '../primitive';
+import { Serializeable } from '../schema';
 
 export function restore(serializedState: Serializeable.GraphSnapshot): GraphSnapshot {
   return new GraphSnapshot();
